+++ src/PauseLayer.cpp
@@ -1,3 +1,4 @@
+#include <Geode/loader/SettingEvent.hpp>
 #include <Geode/modify/PauseLayer.hpp>
 #include <Geode/ui/GeodeUI.hpp>
 #include "Manager.hpp"
@@ -5,9 +6,43 @@
 
 using namespace geode::prelude;
 
+static constexpr float muffleConstant = 2000.f; // undefined set a random number, so will i!
+
+void applyMuffle(Manager *manager = Manager::getSharedInstance()) {
+	manager->system->createDSPByType(FMOD_DSP_TYPE_LOWPASS, &manager->lowPassFilterDSP);
+	manager->lowPassFilterDSP->setParameterFloat(FMOD_DSP_LOWPASS_RESONANCE, 0.f);
+	manager->lowPassFilterDSP->setParameterFloat(FMOD_DSP_LOWPASS_CUTOFF, muffleConstant);
+	manager->channel->addDSP(0, manager->lowPassFilterDSP);
+}
+
+void setVolume(int64_t volumeSetting = Mod::get()->getSettingValue<int64_t>("volume"), double reductionSetting = Mod::get()->getSettingValue<double>("reduction"), Manager *manager = Manager::getSharedInstance()) {
+	manager->channel->setVolume(volumeSetting / 100.f * reductionSetting);
+}
+
+// muffle filter adapted from undefined with consent: https://discord.com/channels/911701438269386882/911702535373475870/1277758859204890738
+
+$execute {
+	listenForSettingChanges<int64_t>("muffleStrength", [](int64_t muffleStrength) {
+		float cutoff = INT_FAST32_MAX;
+		if (muffleStrength != 0) cutoff = muffleConstant / muffleStrength;
+		Manager::getSharedInstance()->lowPassFilterDSP->setParameterFloat(FMOD_DSP_LOWPASS_CUTOFF, cutoff);
+	});
+	listenForSettingChanges<int64_t>("volume", [](int64_t volume) {
+		setVolume(volume, Mod::get()->getSettingValue<double>("reduction"));
+	});
+	listenForSettingChanges<double>("reduction", [](double reduction) {
+		setVolume(Mod::get()->getSettingValue<int64_t>("volume"), reduction);
+	});
+}
+
 class $modify(MyPauseLayer, PauseLayer) {
 	void onModSettings(cocos2d::CCObject* sender) {
 		openSettingsPopup(Mod::get());
+	}
+	void stopMusicRemoveLowPass(Manager *manager = Manager::getSharedInstance()) {
+		manager->channel->removeDSP(manager->lowPassFilterDSP);
+		manager->lowPassFilterDSP->setParameterFloat(FMOD_DSP_LOWPASS_CUTOFF, 0.f);
+		manager->channel->stop();
 	}
 	void customSetup() {
 		PauseLayer::customSetup();
@@ -44,34 +79,36 @@ class $modify(MyPauseLayer, PauseLayer) {
 		manager->system->createSound((manager->path).c_str(), FMOD_LOOP_NORMAL, nullptr, &(manager->sound));
 		manager->sound->setLoopCount(-1);
 		manager->system->playSound((manager->sound), nullptr, false, &(manager->channel));
-		manager->channel->setVolume((Mod::get()->getSettingValue<int64_t>("volume") / 100.f) * Mod::get()->getSettingValue<double>("reduction"));
+		setVolume();
+		if (Mod::get()->getSettingValue<bool>("muffle")) applyMuffle();
 	}
 	void onResume(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onResume(sender);
 	}
 	void onQuit(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onQuit(sender);
 	}
 	void onRestart(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onRestart(sender);
 	}
 	void onRestartFull(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onRestartFull(sender);
 	}
 	void onEdit(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onEdit(sender);
 	}
 	void onNormalMode(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onNormalMode(sender);
 	}
+
 	void onPracticeMode(cocos2d::CCObject* sender) {
-		Manager::getSharedInstance()->channel->stop();
+		stopMusicRemoveLowPass();
 		PauseLayer::onPracticeMode(sender);
 	}
 	void keyDown(cocos2d::enumKeyCodes p0) {