+++ changelog.md
@@ -1,4 +1,8 @@
 # PauseMenuLoop Changelog
+## v1.2.0
+- Added `Muffle Filter` options based on code adapted from Undefined's EditorMusic mod with permission.
+- Allowed settings to be applied to the pause menu loop in real time.
+- Added support for `.flac` and `.oga` files. 
 ## v1.1.0
 - Port to 2.206. Big thanks to CatXus, Aktimoose, ninXout, and hiimjustin000 for helping out with testing!
 ## v1.0.0